requirements.txt
snowflake-connector-python>=3.7.0
pandas>=2.2.0
